Making sure the are no compiler warnings on adhoc code on android

+++ Core/HLE/proAdhoc.cpp
@@ -1328,7 +1328,7 @@ int friendFinder(){
 					else {
 						// Update Ping Time
 						lastping = now;
-						DEBUG_LOG(SCENET, "FriendFinder: Sending OPCODE_PING (%llu)", now);
+						DEBUG_LOG(SCENET, "FriendFinder: Sending OPCODE_PING (%llu)", static_cast<unsigned long long>(now));
 					}
 				}
 			}
